Main class for portrpaths

R6 Reference class encapsulating all portrpaths behaviour and data

Given a yaml formatted local configuration file, not intended to be checked into VCS, the root directory for a given project's dataset may be specified.

This class also provides functionality for changing "profiles" - Swapping between alternate data roots.

For example, given a workflow that analyses some data extracted from some database, with the following structure

- root
  - input
    - patient_history.csv
    - diagnoses.csv
  - output
    - whatever_output.csv

It may be useful/neccesary to extract this data at a reduced scope, before performing a full extract, or perhaps some requirement of this extraction may change.

For example, there is a test subset extract, and then a full extract, having the following structure:

- /media/mnt/SHARED_VOLUME/datasets
  - test_extract
    - input
      - patient_history.csv
      - diagnoses.csv
    - output
      - whatever_output.csv
  - bulk_extract
    - input
      - patient_history.csv
      - diagnoses.csv
    - output
      - whatever_output.csv

PortrPaths allows convenient switching between /media/mnt/SHARED_VOLUME/datasets/bulk_extract & /media/mnt/SHARED_VOLUME/datasets/test_extract.

TODO

Document the input format

Active bindings

profile

Set the active profile

root

Get or set the current root

Methods

Public methods

PortrPath$new()

PortrPath$print()

PortrPath$add_profile()

PortrPath$set_profile()

PortrPath$clone()

Method new()
Initializes new PortrPath object from configuration files given as arguments
Usage

PortrPath$new(local_config_path, shared_config_path = NULL)

Arguments

local_config_path

The location of the configuration file that define local configuration parameters

shared_config_path

DEPRECATED The location of the configuration file shared across the project

Returns
a new PortrPath object

Method print()
Display the PortrPath object
Usage

PortrPath$print()

Method add_profile()
Add a named profile for quickly swapping the root. Profiles allow the user to easily switch between some favourite paths
Usage

PortrPath$add_profile(name, path)

Arguments

name

the friendly name to refer to the profile

path

the path the profile should refer to

Returns
None

Method set_profile()
Alias of path$profile <- name
Usage

PortrPath$set_profile(name)

Arguments

name

the name of the profile to switch to

Method clone()
The objects of this class are cloneable with this method.
Usage

PortrPath$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.